backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
torch==2.1.0
transformers==4.36.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
opencv-python==4.8.1.78
pillow==10.1.0
nltk==3.8.1
spacy==3.7.2
pytesseract==0.3.10
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
